Final fixes

+++ app/Service/CategoryService.php
@@ -6,6 +6,7 @@ namespace App\Service;
 use App\Category;
 
 use App\Repository\CategoryRepository;
+use App\Repository\ProductRepository;
 use Illuminate\Validation\ValidationException;
 
 
@@ -15,15 +16,19 @@ class CategoryService
      * @var CategoryRepository
      */
     private $categoryRepository;
+    private $productRepository;
 
     /**
      * CategoryService constructor.
      * @param CategoryRepository $categoryRepository
+     * @param ProductRepository $productRepository
      */
-    public function __construct(CategoryRepository $categoryRepository)
+    public function __construct(CategoryRepository $categoryRepository, ProductRepository $productRepository)
     {
         $this->categoryRepository = $categoryRepository;
+        $this->productRepository = $productRepository;
     }
+
 
     public function create(string $name)
     {
@@ -37,7 +42,14 @@ class CategoryService
     {
         $category = $this->categoryRepository->findOrfail($id);
 
+        $productCategory = $this->productRepository->findByCategory($category);
+
+        if (!empty($productCategory->toArray())) {
+            throw ValidationException::withMessages(['category' => 'No se puede eliminar porque existen productos con la categoría']);
+        }
+
         $this->categoryRepository->delete($category);
+
     }
 
     public function edit (string $id, string $name)
